Removed some unused lines
Trying to fix the error...
The page build failed for the `master` branch with the following error:

There was a YAML syntax error on line 32 column 7 in `_data/data.yml`: `did not find expected '-' indicator while parsing a block collection`. For more information, see https://help.github.com/en/github/working-with-github-pages/troubleshooting-jekyll-build-errors-for-github-pages-sites#invalid-yaml-in-data-file.